devtools/server/tests/chrome/test_framerate_01.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1007200 - Create a framerate actor
-->
<head>
  <meta charset="utf-8">
  <title>Framerate actor test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="framerate-helpers.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script>
"use strict";

window.onload = async function() {
  const target = await getTargetForSelectedTab();
  const front = await target.getFront("framerate");
  const TICK = 1000;

  await waitFor(TICK);
  await front.startRecording();
  await waitFor(TICK);
  const rawData = await front.stopRecording();
  await onRecordingStopped(front, rawData);
  await target.destroy();
  SimpleTest.finish();
};

// Local Helpers
async function onRecordingStopped(front, rawData) {
  ok(rawData, "There should be a recording available.");

  const timeline = plotFPS(rawData);
  ok(timeline.length >= 2,
    "There should be at least one measurement available, with two entries.");

  let prevTimeStart = timeline[0].delta;

  for (let i = 0; i < timeline.length; i += 2) {
    const currTimeStart = timeline[i].delta;
    const currTimeEnd = timeline[i + 1].delta;
    info("Testing delta: " + currTimeStart + " vs. " + currTimeEnd);

    ok(currTimeStart < currTimeEnd,
      "The start and end time deltas should be consecutive.");
    is(currTimeStart, prevTimeStart,
      "There should be two time deltas for each framerate value.");

    prevTimeStart = currTimeEnd;
  }

  for (let i = 0; i < timeline.length; i += 2) {
    const currFramerateStart = timeline[i].value;
    const currFramerateEnd = timeline[i + 1].value;
    info("Testing framerate: " + currFramerateStart);

    is(currFramerateStart, currFramerateEnd,
      "The start and end framerate values should be equal.");

    is(typeof currFramerateStart, "number", "All values should be numbers.");
    ok(currFramerateStart <= 60, "All values were correctly clamped.");
  }
}
</script>
</pre>
</body>
</html>
